[REFACTOR/GROUP]: 그룹 하위 패키지 추가
