+++ src/main/java/com/wable/harmonika/domain/group/dto/UserRoleUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.wable.harmonika.domain.group;
+package com.wable.harmonika.domain.group.dto;
 
 import com.wable.harmonika.global.entity.Role;
 import lombok.Getter;
