+++ src/main/java/com/wable/harmonika/domain/group/repository/GroupQuestionRepository.java
@@ -1,4 +1,4 @@
-package com.wable.harmonika.domain.group;
+package com.wable.harmonika.domain.group.repository;
 
 import com.wable.harmonika.domain.group.entity.GroupQuestion;
 import com.wable.harmonika.domain.group.entity.Groups;
