+++ src/main/java/com/wable/harmonika/domain/group/repository/UserGroupRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.wable.harmonika.domain.group;
+package com.wable.harmonika.domain.group.repository;
 
 import com.wable.harmonika.domain.user.entity.Users;
 import java.util.List;
