+++ src/main/java/com/wable/harmonika/domain/profile/service/ProfileService.java
@@ -1,6 +1,6 @@
 package com.wable.harmonika.domain.profile.service;
 
-import com.wable.harmonika.domain.group.GroupRepository;
+import com.wable.harmonika.domain.group.repository.GroupRepository;
 import com.wable.harmonika.domain.profile.dto.CreateProfileByGroupDto;
 import com.wable.harmonika.domain.profile.dto.CreateProfileByUserDto;
 import com.wable.harmonika.domain.profile.dto.QuestionDataDto;
@@ -16,7 +16,6 @@ import com.wable.harmonika.global.error.exception.InvalidException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import net.minidev.asm.ConvertDate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.beans.factory.annotation.Value;
@@ -25,7 +24,6 @@ import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 import software.amazon.awssdk.services.s3.presigner.model.PutObjectPresignRequest;
 
 import java.time.Duration;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
